requirements.txt
whisperx
PySide6
pygame
sounddevice
python-dotenv
ffmpeg-python
llm
chromadb